Implement experimental NEON support for FIR filter in fdmdv RX
The FIR filter consumes much of the CPU in RX mode on ARM platforms.
This is a significant performance concern for smaller boxes such
as the RPi and the FLEX6k series. This commit implements the FIR
filter using NEON GCC intrinsics to hopefully accomplish a speed
up on these platforms.  Note that you *must* use the -mfpu=neon
flag for this to activate.